# postman/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MCQ Online Test Platform API",
...
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/examiner/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Examiner login",
        "description": "Authenticate an examiner with username and password.\n\nOn success an HttpOnly `session_id` cookie (1 hour TTL) is set on the\nresponse. After 5 failed attempts within 15 minutes the account is\ntemporarily locked (423) and `locked_until` indicates when it reopens.",
        "operationId": "examiner_login_api_v1_auth_examiner_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExaminerLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Examiner profile; the session cookie is set on the response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_ExaminerLoginResponse_"
                }
              }
            }
          },
          "400": {
            "description": "Validation error \u2014 one or more request fields are invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/validation-error",
                  "title": "Validation Error",
                  "status": 400,
                  "detail": "One or more fields failed validation.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b",
                  "errors": [
                    {
                      "field": "username",
                      "message": "String should have at least 1 character"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Invalid username or password.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/invalid-credentials",
                  "title": "Invalid Credentials",
                  "status": 401,
                  "detail": "Invalid username or password",
                  "instance": "/api/v1/auth/examiner/login",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "423": {
            "description": "Account locked after too many failed login attempts.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/locked",
                  "title": "Account Locked",
                  "status": 423,
                  "detail": "This account is temporarily locked due to too many failed login attempts.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b",
                  "locked_until": "2026-01-01T12:30:00Z"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/examinee/entry": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Examinee entry (register or resume)",
        "description": "Register a new examinee or resume an in-progress test session.\n\nIdentified by PRN (registration number) and name; privacy acknowledgement\nis mandatory. Returns **201** when a new account and session are created,\n**200** when an existing in-progress session is resumed. An HttpOnly\n`session_id` cookie (4 hour TTL) is set on the response. Registration is\nrate-limited per IP (429).",
        "operationId": "examinee_entry_api_v1_auth_examinee_entry_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamineeEntryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The active test session for the examinee (existing session resumed).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__schemas__common__SuccessResponse_ExamineeSessionResource___1"
                }
              }
            }
          },
          "201": {
            "description": "A new examinee account and test session were created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__schemas__common__SuccessResponse_ExamineeSessionResource___1"
                }
              }
            }
          },
          "400": {
            "description": "Validation error \u2014 one or more request fields are invalid.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/validation-error",
                  "title": "Validation Error",
                  "status": 400,
                  "detail": "One or more fields failed validation.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b",
                  "errors": [
                    {
                      "field": "username",
                      "message": "String should have at least 1 character"
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Identity mismatch \u2014 the PRN exists but the name does not match.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/identity-mismatch",
                  "title": "Identity Mismatch",
                  "status": 403,
                  "detail": "The provided details do not match our records.",
                  "instance": "/api/v1/auth/examinee/entry",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "409": {
            "description": "This registration number has already completed the test (retakes disabled).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/exam-already-completed",
                  "title": "Exam Already Completed",
                  "status": 409,
                  "detail": "This registration number has already completed the test.",
                  "instance": "/api/v1/auth/examinee/entry",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded \u2014 retry after the indicated delay.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/rate-limit-exceeded",
                  "title": "Rate Limit Exceeded",
                  "status": 429,
                  "detail": "Too many requests. Please try again later.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable \u2014 a required dependency or resource is not ready.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/service-unavailable",
                  "title": "Insufficient Question Bank",
                  "status": 503,
                  "detail": "The question bank does not have enough active questions.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Log out",
        "description": "Invalidate the current session server-side and clear the session cookie.\n\nRequires an authenticated session and the `X-CSRF-Token` header.",
        "operationId": "logout_api_v1_auth_logout_post",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Session invalidated; the session cookie is cleared."
          },
          "401": {
            "description": "Unauthenticated \u2014 missing or expired session cookie.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/unauthenticated",
                  "title": "Unauthenticated",
                  "status": 401,
                  "detail": "Authentication required.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden \u2014 insufficient role or CSRF token validation failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/forbidden",
                  "title": "Forbidden",
                  "status": 403,
                  "detail": "You do not have permission to perform this action.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": [],
            "csrfToken": []
          }
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Get current user profile",
        "description": "Return the authenticated user's profile and session-bound CSRF token.\n\nRole-specific fields are populated: `username`/`is_admin` for examiners,\n`prn`/`name` for examinees. Use `csrf_token` in the `X-CSRF-Token` header\non subsequent mutating requests.",
        "operationId": "get_current_user_profile_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Profile of the authenticated user, including the session CSRF token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_CurrentUserResponse_"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated \u2014 missing or expired session cookie.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://api.example.com/problems/unauthenticated",
                  "title": "Unauthenticated",
                  "status": 401,
                  "detail": "Authentication required.",
                  "instance": "/api/v1/...",
                  "request_id": "9f1b2c3d4e5f6a7b"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExamineeEntryRequest": {
        "properties": {
          "prn": {
            "type": "string",
            "maxLength": 20,
            "minLength": 1,
            "title": "Prn",
            "description": "Permanent registration number (normalised to uppercase).",
            "examples": [
              "PRN2026001"
            ]
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "description": "Examinee's full name as registered (whitespace trimmed).",
            "examples": [
              "Jane Doe"
            ]
          },
          "privacy_acknowledged": {
            "type": "boolean",
            "title": "Privacy Acknowledged",
            "description": "Must be true; confirms the examinee accepted the privacy notice.",
            "examples": [
              true
            ]
          }
        },
        "type": "object",
        "required": [
          "prn",
          "name",
          "privacy_acknowledged"
        ],
        "title": "ExamineeEntryRequest",
        "description": "Examinee identification for registration or session resumption."
      },
...
      "ExaminerLoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Username",
            "description": "Examiner username (leading/trailing whitespace is trimmed).",
            "examples": [
              "examiner1"
            ]
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "description": "Examiner password.",
            "examples": [
              "correct horse battery staple"
            ]
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "ExaminerLoginRequest",
        "description": "Credentials for examiner login."
      },
      "ExaminerLoginResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id",
            "description": "Unique identifier of the examiner."
          },
          "username": {
            "type": "string",
            "title": "Username",
            "description": "Examiner username.",
            "examples": [
              "examiner1"
            ]
          },
          "must_change_password": {
            "type": "boolean",
            "title": "Must Change Password",
            "description": "True when the examiner must change their password before continuing.",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "username"
        ],
        "title": "ExaminerLoginResponse",
        "description": "Examiner profile returned after a successful login."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResponseMeta": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "Unique request identifier; matches the X-Request-Id response header.",
            "examples": [
              "9f1b2c3d4e5f6a7b"
            ]
          }
        },
        "type": "object",
        "required": [
          "request_id"
        ],
        "title": "ResponseMeta",
        "description": "Per-request metadata included in every success envelope."
      },
...
      "SessionStatus": {
        "type": "string",
        "enum": [
          "ACTIVE",
          "COMPLETED",
          "EXPIRED"
        ],
        "title": "SessionStatus"
      },
...
      "SuccessResponse_CurrentUserResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CurrentUserResponse",
            "description": "The endpoint-specific response payload."
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta",
            "description": "Request metadata for log correlation."
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "SuccessResponse[CurrentUserResponse]"
      },
...
      "SuccessResponse_ExaminerLoginResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ExaminerLoginResponse",
            "description": "The endpoint-specific response payload."
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta",
            "description": "Request metadata for log correlation."
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "SuccessResponse[ExaminerLoginResponse]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "app__schemas__auth__ExamineeSessionResource": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id",
            "description": "Unique identifier of the test session."
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus",
            "description": "Current session status."
          },
          "total_questions": {
            "type": "integer",
            "title": "Total Questions",
            "description": "Number of questions in the test.",
            "default": 50
          },
          "answered_count": {
            "type": "integer",
            "title": "Answered Count",
            "description": "Number of questions answered so far."
          },
          "current_position": {
            "type": "integer",
            "title": "Current Position",
            "description": "Position (1-based) of the question the examinee is on.",
            "default": 1
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At",
            "description": "When the session was started (UTC)."
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At",
            "description": "When the session expires (UTC)."
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "status",
          "answered_count",
          "started_at",
          "expires_at"
        ],
        "title": "ExamineeSessionResource",
        "description": "Summary of an examinee's test session."
      },
      "app__schemas__common__SuccessResponse_ExamineeSessionResource___1": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/app__schemas__auth__ExamineeSessionResource",
            "description": "The endpoint-specific response payload."
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta",
            "description": "Request metadata for log correlation."
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "SuccessResponse[ExamineeSessionResource]"
      },
...
      "FieldError": {
        "description": "A single field-level validation failure.",
        "properties": {
          "field": {
            "description": "Dotted path of the invalid request field.",
            "examples": [
              "username"
            ],
            "title": "Field",
            "type": "string"
          },
          "message": {
            "description": "Human-readable description of the validation failure.",
            "examples": [
              "String should have at least 1 character"
            ],
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "title": "FieldError",
        "type": "object"
      },
      "ProblemDetails": {
        "description": "RFC 7807 problem details returned with content type application/problem+json.",
        "properties": {
          "type": {
            "description": "URI identifying the problem type.",
            "examples": [
              "https://api.example.com/problems/unauthenticated"
            ],
            "title": "Type",
            "type": "string"
          },
          "title": {
            "description": "Short, human-readable summary of the problem type.",
            "examples": [
              "Unauthenticated"
            ],
            "title": "Title",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code.",
            "examples": [
              401
            ],
            "title": "Status",
            "type": "integer"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Human-readable explanation specific to this occurrence.",
            "examples": [
              "Authentication required."
            ],
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Request path on which the problem occurred.",
            "examples": [
              "/api/v1/auth/me"
            ],
            "title": "Instance"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Request identifier for log correlation; matches X-Request-Id.",
            "examples": [
              "9f1b2c3d4e5f6a7b"
            ],
            "title": "Request Id"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Field-level validation errors (only on 400 validation failures).",
            "title": "Errors"
          },
          "locked_until": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "When a locked account becomes available again (only on 423 responses).",
            "title": "Locked Until"
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ],
        "title": "ProblemDetails",
        "type": "object"
      }
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session_id",
        "description": "Opaque server-side session token issued by the login/entry endpoints as an HttpOnly cookie. Examiner sessions last 1 hour, examinee sessions 4 hours by default."
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Session-bound CSRF token required on mutating requests. Obtain it from GET /auth/me (or GET /auth/csrf before login)."
      }
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ]
}